+++ Lignite Music/LMMusicCollectionsView.m
@@ -47,7 +47,7 @@
 	}
 }
 
-- (float)contentSubviewFactorial:(BOOL)height forBigListEntry:(LMBigListEntry *)bigListEntry {
+- (CGFloat)contentSubviewFactorial:(BOOL)height forBigListEntry:(LMBigListEntry *)bigListEntry {
 	CGRect frame = [[self.collectionView.visibleCells firstObject] frame];
 	
 	if(frame.size.width == 0){
@@ -60,7 +60,7 @@
 	return height ? 0.1 : ([LMLayoutManager isExtraSmall] ? 0.9 : 1.0);
 }
 
-- (void)sizeChangedToLargeSize:(BOOL)largeSize withHeight:(float)newHeight forBigListEntry:(LMBigListEntry*)bigListEntry {
+- (void)sizeChangedToLargeSize:(BOOL)largeSize withHeight:(CGFloat)newHeight forBigListEntry:(LMBigListEntry*)bigListEntry {
 	//If the new size is large/opened
 	//	NSLog(@"%@ changed large", bigListEntry);
 }
